:tractor: relocated the pip_install code to deprecated folder. ready for future clean up
