+++ moban/mobanfile/__init__.py
@@ -5,8 +5,8 @@ from collections import OrderedDict
 
 from moban import plugins, reporter, constants, file_system
 from lml.utils import do_import
-from moban.utils import merge, pip_install, verify_the_existence_of_directories
-from moban.deprecated import GitRequire, deprecated
+from moban.utils import merge, verify_the_existence_of_directories
+from moban.deprecated import GitRequire, deprecated, pip_install
 from moban.deprecated.repo import git_clone
 from moban.plugins.template import expand_template_directories
 from moban.mobanfile.targets import parse_targets, extract_group_targets
