+++ tests/test_utils.py
@@ -15,7 +15,7 @@ def test_mkdir_p():
 @patch("subprocess.check_call")
 def test_pip_install(fake_check_all):
     import sys
-    from moban.utils import pip_install
+    from moban.deprecated import pip_install
 
     pip_install(["package1", "package2"])
     fake_check_all.assert_called_with(
